Move test files to new folder
