+++ src/components/__tests__/CreditCardExpirationDateInput.test.tsx
@@ -2,17 +2,17 @@ import * as React from 'react'
 import { mount } from 'enzyme'
 import { expect } from 'chai'
 
-import CreditCardPostalCodeInput from './CreditCardPostalCodeInput'
-import { ContextInterface } from './Context'
+import CreditCardExpirationDateInput from '../CreditCardExpirationDateInput'
+import { ContextInterface } from '../Context'
 
 let mockContext: ContextInterface = {}
-jest.mock('./Context', () => ({
+jest.mock('../Context', () => ({
   ContextConsumer: (props: { children: (context: ContextInterface) => {} }) => {
     return props.children(mockContext)
   },
 }))
 
-describe('CreditCardPostalCodeInput', () => {
+describe('CreditCardExpirationDateInput', () => {
   beforeEach(() => {
     mockContext = {}
   })
@@ -20,30 +20,30 @@ describe('CreditCardPostalCodeInput', () => {
   describe('DOM attributes', () => {
     it('should render the form id placeholder', () => {
       mockContext = { formId: 'my-form' }
-      const wrapper = mount(<CreditCardPostalCodeInput />)
-      expect(wrapper.find(`#${mockContext.formId}-sq-postal-code`)).to.be.length(1)
+      const wrapper = mount(<CreditCardExpirationDateInput />)
+      expect(wrapper.find(`#${mockContext.formId}-sq-expiration-date`)).to.be.length(1)
     })
 
     it('should render the class name for styling', () => {
-      const wrapper = mount(<CreditCardPostalCodeInput />)
+      const wrapper = mount(<CreditCardExpirationDateInput />)
       expect(wrapper.find('.sq-label')).to.be.length(1)
     })
   })
 
   describe('label', () => {
     it('should render the default placeholder label', () => {
-      const wrapper = mount(<CreditCardPostalCodeInput />)
-      expect(wrapper.find('.sq-label').text()).to.eql('Postal')
+      const wrapper = mount(<CreditCardExpirationDateInput />)
+      expect(wrapper.find('.sq-label').text()).to.eql('Expiration')
     })
 
     it('should render a custom placeholder label', () => {
       const label = 'test'
-      const wrapper = mount(<CreditCardPostalCodeInput label={label} />)
+      const wrapper = mount(<CreditCardExpirationDateInput label={label} />)
       expect(wrapper.find('.sq-label').text()).to.eql(label)
     })
 
     it('should not render the placeholder label', () => {
-      const wrapper = mount(<CreditCardPostalCodeInput label={''} />)
+      const wrapper = mount(<CreditCardExpirationDateInput label={''} />)
       expect(wrapper.find('.sq-label')).to.be.length(0)
     })
   })
