+++ src/components/__tests__/CreditCardNumberInput.test.tsx
@@ -2,17 +2,17 @@ import * as React from 'react'
 import { mount } from 'enzyme'
 import { expect } from 'chai'
 
-import CreditCardCVVInput from './CreditCardCVVInput'
-import { ContextInterface } from './Context'
+import CreditCardNumberInput from '../CreditCardNumberInput'
+import { ContextInterface } from '../Context'
 
 let mockContext: ContextInterface = {}
-jest.mock('./Context', () => ({
+jest.mock('../Context', () => ({
   ContextConsumer: (props: { children: (context: ContextInterface) => {} }) => {
     return props.children(mockContext)
   },
 }))
 
-describe('CreditCardCVVInput', () => {
+describe('CreditCardNumberInput', () => {
   beforeEach(() => {
     mockContext = {}
   })
@@ -20,30 +20,30 @@ describe('CreditCardCVVInput', () => {
   describe('DOM attributes', () => {
     it('should render the form id placeholder', () => {
       mockContext = { formId: 'my-form' }
-      const wrapper = mount(<CreditCardCVVInput />)
-      expect(wrapper.find(`#${mockContext.formId}-sq-cvv`)).to.be.length(1)
+      const wrapper = mount(<CreditCardNumberInput />)
+      expect(wrapper.find(`#${mockContext.formId}-sq-card-number`)).to.be.length(1)
     })
 
     it('should render the class name for styling', () => {
-      const wrapper = mount(<CreditCardCVVInput />)
+      const wrapper = mount(<CreditCardNumberInput />)
       expect(wrapper.find('.sq-label')).to.be.length(1)
     })
   })
 
   describe('label', () => {
     it('should render the default placeholder label', () => {
-      const wrapper = mount(<CreditCardCVVInput />)
-      expect(wrapper.find('.sq-label').text()).to.eql('CVV')
+      const wrapper = mount(<CreditCardNumberInput />)
+      expect(wrapper.find('.sq-label').text()).to.eql('Credit Card')
     })
 
     it('should render a custom placeholder label', () => {
       const label = 'test'
-      const wrapper = mount(<CreditCardCVVInput label={label} />)
+      const wrapper = mount(<CreditCardNumberInput label={label} />)
       expect(wrapper.find('.sq-label').text()).to.eql(label)
     })
 
     it('should not render the placeholder label', () => {
-      const wrapper = mount(<CreditCardCVVInput label={''} />)
+      const wrapper = mount(<CreditCardNumberInput label={''} />)
       expect(wrapper.find('.sq-label')).to.be.length(0)
     })
   })
