+++ src/components/__tests__/CreditCardPostalCodeInput.test.tsx
@@ -2,17 +2,17 @@ import * as React from 'react'
 import { mount } from 'enzyme'
 import { expect } from 'chai'
 
-import CreditCardNumberInput from './CreditCardNumberInput'
-import { ContextInterface } from './Context'
+import CreditCardPostalCodeInput from '../CreditCardPostalCodeInput'
+import { ContextInterface } from '../Context'
 
 let mockContext: ContextInterface = {}
-jest.mock('./Context', () => ({
+jest.mock('../Context', () => ({
   ContextConsumer: (props: { children: (context: ContextInterface) => {} }) => {
     return props.children(mockContext)
   },
 }))
 
-describe('CreditCardNumberInput', () => {
+describe('CreditCardPostalCodeInput', () => {
   beforeEach(() => {
     mockContext = {}
   })
@@ -20,30 +20,30 @@ describe('CreditCardNumberInput', () => {
   describe('DOM attributes', () => {
     it('should render the form id placeholder', () => {
       mockContext = { formId: 'my-form' }
-      const wrapper = mount(<CreditCardNumberInput />)
-      expect(wrapper.find(`#${mockContext.formId}-sq-card-number`)).to.be.length(1)
+      const wrapper = mount(<CreditCardPostalCodeInput />)
+      expect(wrapper.find(`#${mockContext.formId}-sq-postal-code`)).to.be.length(1)
     })
 
     it('should render the class name for styling', () => {
-      const wrapper = mount(<CreditCardNumberInput />)
+      const wrapper = mount(<CreditCardPostalCodeInput />)
       expect(wrapper.find('.sq-label')).to.be.length(1)
     })
   })
 
   describe('label', () => {
     it('should render the default placeholder label', () => {
-      const wrapper = mount(<CreditCardNumberInput />)
-      expect(wrapper.find('.sq-label').text()).to.eql('Credit Card')
+      const wrapper = mount(<CreditCardPostalCodeInput />)
+      expect(wrapper.find('.sq-label').text()).to.eql('Postal')
     })
 
     it('should render a custom placeholder label', () => {
       const label = 'test'
-      const wrapper = mount(<CreditCardNumberInput label={label} />)
+      const wrapper = mount(<CreditCardPostalCodeInput label={label} />)
       expect(wrapper.find('.sq-label').text()).to.eql(label)
     })
 
     it('should not render the placeholder label', () => {
-      const wrapper = mount(<CreditCardNumberInput label={''} />)
+      const wrapper = mount(<CreditCardPostalCodeInput label={''} />)
       expect(wrapper.find('.sq-label')).to.be.length(0)
     })
   })
